pyproject.toml
[tool.poetry]
name = "tradingBot"
version = "0.1.0"
description = ""
authors = ["drlux <sorrentino.luca@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.2"
hydra-core = "^1.2.0"
PyYAML = "^6.0"
pathlib = "^1.0.1"
requests = "^2.27.1"
pandas = "^1.4.2"
plotly = "^5.8.0"
colorama = "^0.4.4"
telegram-send = "^0.34"
Pillow = "^9.1.1"
kaleido = "0.2.1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
hydra-core = "^1.2.0"
PyYAML = "^6.0"
pathlib = "^1.0.1"
requests = "^2.27.1"
pandas = "^1.4.2"
plotly = "^5.8.0"
Pillow = "^9.1.1"
kaleido = "^0.2.1"
traitlets = "^5.2.1"
pickleshare = "^0.7.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
